+++ src/class/plugins/class-plugins-detect.php
@@ -113,7 +113,7 @@ class Plugins_Detect {
 	 * @return bool true if AMP plugin is active
 	 */
 	private function amp_is_active(): bool {
-		return $this->lumiere_is_amp_page(); // Trait Main.
+		return $this->is_amp_page(); // Trait Main.
 	}
 
 	/**